+++ src/whatsapp/controllers/instance.controller.ts
@@ -64,7 +64,10 @@ export class InstanceController {
         this.repository,
         this.cache,
       );
-      instance.instanceName = instanceName;
+      instance.instanceName = instanceName
+        .toLowerCase()
+        .replace(/[^a-z0-9]/g, '')
+        .replace(' ', '');
       this.logger.verbose('instance: ' + instance.instanceName + ' created');
 
       this.waMonitor.waInstances[instance.instanceName] = instance;
@@ -191,7 +194,10 @@ export class InstanceController {
         this.repository,
         this.cache,
       );
-      instance.instanceName = instanceName;
+      instance.instanceName = instanceName
+        .toLowerCase()
+        .replace(/[^a-z0-9]/g, '')
+        .replace(' ', '');
 
       this.logger.verbose('instance: ' + instance.instanceName + ' created');
 
@@ -357,24 +363,8 @@ export class InstanceController {
     try {
       this.logger.verbose('requested restartInstance from ' + instanceName + ' instance');
 
-      this.logger.verbose('deleting instance: ' + instanceName);
-      delete this.waMonitor.waInstances[instanceName];
-
-      this.logger.verbose('creating instance: ' + instanceName);
-      const instance = new WAStartupService(
-        this.configService,
-        this.eventEmitter,
-        this.repository,
-        this.cache,
-      );
-
-      instance.instanceName = instanceName;
-
-      this.logger.verbose('instance: ' + instance.instanceName + ' created');
-
-      this.logger.verbose('connecting instance: ' + instanceName);
-      await instance.connectToWhatsapp();
-      this.waMonitor.waInstances[instance.instanceName] = instance;
+      this.logger.verbose('logging out instance: ' + instanceName);
+      this.waMonitor.waInstances[instanceName]?.client?.ws?.close();
 
       return { error: false, message: 'Instance restarted' };
     } catch (error) {